requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
Pillow==6.1.0
pkg-resources==0.0.0
six==1.12.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
Werkzeug==0.15.6
